+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [1.6.45] - 2017-03-22
+
+### Added
+
+-  Added `HeimrichHannot\Haste\Util\Validator` class with `IBAN` and `BIC` validation
+- `globalcitizen/php-iban` dependency added 
+- added rgxp `iban` to validate input values
+- added rgxp `bic` to validate input values
+
 ## [1.6.44] - 2017-03-22
 
 ### Added